+++ CHANGELOG.md
@@ -27,6 +27,8 @@ an external audit database by using config.audit_database configuration entry.
 
 - API endpoints for granting and revoking role membership
 
+- API endpoint for the role graph
+
 ## [0.4.0] - 2018-04-10
 ### Added
 - Policy changes now generate audit log messages. These can optionally be generated in RFC5424